Add files via upload
Since an update of libraries special characters are no longer read in correctly the the command read.systat. On a computer with older library versions the command works correctly. It is, however, not the foreign version which causes the problem (I replaed 0.8-84 with 0.8-81 which is the same as on the older computer) but probably the libraries loaded at start of R. Is it possible to reset these?